Move analytics utility functions to a separate file
